added postfix impl, redesigned main and io funcs with usage new postfix class
